requirements.txt
tweepy==4.15.0
python-dotenv==1.0.1
Flask==3.0.2
gunicorn==21.2.0 